+++ rust/headless-client/src/linux.rs
@@ -14,8 +14,6 @@ pub(crate) fn default_token_path() -> PathBuf {
     PathBuf::from("/etc").join(BUNDLE_ID).join("token")
 }
 
-pub use firezone_bin_shared::linux::{tcp_socket_factory, udp_socket_factory};
-
 pub(crate) fn check_token_permissions(path: &Path) -> Result<()> {
     let Ok(stat) = nix::sys::stat::fstatat(None, path, nix::fcntl::AtFlags::empty()) else {
         // File doesn't exist or can't be read
